Rename ConveringTests -> Margin
