+++ Hosts/VS/CommonUI/CommonUI.csproj
@@ -116,14 +116,15 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CoveringTests\MainUserControl1.xaml.cs">
-      <DependentUpon>MainUserControl1.xaml</DependentUpon>
+    <Compile Include="Margin\MainUserControl.xaml.cs">
+      <DependentUpon>MainUserControl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="CoveringTests\ViewModel\CoveringTestViewModel.cs" />
-    <Compile Include="CoveringTests\ViewModel\MainViewModel.GlyphInfo.cs" />
-    <Compile Include="CoveringTests\ViewModel\MainViewModel.HostIdeActions.cs" />
-    <Compile Include="CoveringTests\ViewModel\MainViewModel1.cs" />
-    <Compile Include="CoveringTests\ViewModel\ViewModelLocator.cs" />
+    <Compile Include="Margin\ViewModel\CoveringTestViewModel.cs" />
+    <Compile Include="Margin\ViewModel\MainViewModel.cs" />
+    <Compile Include="Margin\ViewModel\MainViewModel.HostIdeActions.cs">
+      <DependentUpon>MainViewModel.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Margin\ViewModel\ViewModelLocator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StatusBar\RunStateToIconTextConverter.cs" />
     <Compile Include="StatusBar\RunStateToIconColorConverter.cs" />
@@ -142,7 +143,7 @@
     <Compile Include="TestCode\FakeWpfTextView.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Page Include="CoveringTests\MainUserControl1.xaml">
+    <Page Include="Margin\MainUserControl.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -158,18 +159,18 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="CoveringTests\Resources\actiongototest.png" />
-    <Resource Include="CoveringTests\Resources\actionstartwithdebug.png" />
-    <Resource Include="CoveringTests\Resources\actionstartwithoutdebug.png" />
-    <Resource Include="CoveringTests\Resources\testresultfailed.png" />
-    <Resource Include="CoveringTests\Resources\testresultpassed.png" />
-    <Resource Include="CoveringTests\Resources\testresultunknown.png" />
+    <Resource Include="Margin\Resources\actiongototest.png" />
+    <Resource Include="Margin\Resources\actionstartwithdebug.png" />
+    <Resource Include="Margin\Resources\actionstartwithoutdebug.png" />
+    <Resource Include="Margin\Resources\testresultfailed.png" />
+    <Resource Include="Margin\Resources\testresultpassed.png" />
+    <Resource Include="Margin\Resources\testresultunknown.png" />
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="CoveringTests\Resources\actionexpand.png" />
+    <Resource Include="Margin\Resources\actionexpand.png" />
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="CoveringTests\Resources\actioncollapse.png" />
+    <Resource Include="Margin\Resources\actioncollapse.png" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\..\..\packages\Microsoft.Diagnostics.Tracing.EventRegister.1.0.26\build\Microsoft.Diagnostics.Tracing.EventRegister.targets" Condition="('$(Configuration)' != 'Debug') and Exists('..\..\..\packages\Microsoft.Diagnostics.Tracing.EventRegister.1.0.26\build\Microsoft.Diagnostics.Tracing.EventRegister.targets')" />
