--- Hosts/VS/CommonUI/CoveringTests/ViewModel/MainViewModel.GlyphInfo.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows.Media;
-
-namespace R4nd0mApps.TddStud10.Hosts.Common.CoveringTests.ViewModel
-{
-    public class GlyphInfo
-    {
-        public Geometry Shape { get; set; }
-        public Color Color { get; set; }
-        public double OutlineThickness { get; set; }
-    }
-}
