+++ Hosts/VS/CommonUI/Margin/MainUserControl.xaml
@@ -1,10 +1,10 @@
 ﻿<UserControl 
-    x:Class="R4nd0mApps.TddStud10.Hosts.Common.CoveringTests.MainUserControl1"
+    x:Class="R4nd0mApps.TddStud10.Hosts.Common.Margin.MainUserControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-    xmlns:local="clr-namespace:R4nd0mApps.TddStud10.Hosts.Common.CoveringTests"
+    xmlns:local="clr-namespace:R4nd0mApps.TddStud10.Hosts.Common.Margin"
     xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
     xmlns:mvvm="http://www.galasoft.ch/mvvmlight"
 >
