+++ Hosts/VS/CommonUI/Margin/MainUserControl.xaml.cs
@@ -1,18 +1,18 @@
 ﻿using System.Windows.Controls;
 
-namespace R4nd0mApps.TddStud10.Hosts.Common.CoveringTests
+namespace R4nd0mApps.TddStud10.Hosts.Common.Margin
 {
     /// <summary>
     /// Interaction logic for UserControl.xaml
     /// </summary>
-    public partial class MainUserControl1 : UserControl
+    public partial class MainUserControl : UserControl
     {
         public Canvas Canvas
         {
             get { return this.canvas; }
         }
 
-        public MainUserControl1()
+        public MainUserControl()
         {
             InitializeComponent();
         }
