+++ Hosts/VS/CommonUI/Margin/Resources/testresultpassed.png
