+++ Hosts/VS/CommonUI/Margin/ViewModel/MainViewModel.HostIdeActions.cs
@@ -3,7 +3,7 @@ using R4nd0mApps.TddStud10.Common.Domain;
 using System;
 using System.Collections.Generic;
 
-namespace R4nd0mApps.TddStud10.Hosts.Common.CoveringTests.ViewModel
+namespace R4nd0mApps.TddStud10.Hosts.Common.Margin.ViewModel
 {
     public class HostIdeActions
     {
