+++ Hosts/VS/CommonUI/Margin/ViewModel/ViewModelLocator.cs
@@ -1,7 +1,7 @@
 ﻿using GalaSoft.MvvmLight.Ioc;
 using Microsoft.Practices.ServiceLocation;
 
-namespace R4nd0mApps.TddStud10.Hosts.Common.CoveringTests.ViewModel
+namespace R4nd0mApps.TddStud10.Hosts.Common.Margin.ViewModel
 {
     public class ViewModelLocator
     {
@@ -24,14 +24,14 @@ namespace R4nd0mApps.TddStud10.Hosts.Common.CoveringTests.ViewModel
             ////}
 
             //SimpleIoc.Default.Register<IDataAccessService, DataAccessService>();
-            SimpleIoc.Default.Register<MainViewModel1>();
+            SimpleIoc.Default.Register<MainViewModel>();
         }
 
-        public MainViewModel1 Main
+        public MainViewModel Main
         {
             get
             {
-                return ServiceLocator.Current.GetInstance<MainViewModel1>();
+                return ServiceLocator.Current.GetInstance<MainViewModel>();
             }
         }
