+++ Hosts/VS/TddStudioPackage.Core/Editor/MarginCanvas.fs
@@ -2,14 +2,14 @@
 
 open System.Windows.Controls
 open System.Windows
-open R4nd0mApps.TddStud10.Hosts.Common.CoveringTests
-open R4nd0mApps.TddStud10.Hosts.Common.CoveringTests.ViewModel
+open R4nd0mApps.TddStud10.Hosts.Common.Margin
+open R4nd0mApps.TddStud10.Hosts.Common.Margin.ViewModel
 
 
 // NOTE: This class should not contain any business other than the l/t/w/h set + seq -> UIElementCollction copy.
 // Hence this class is not covered by unit tests  
 type MarginCanvas(getZL) = 
-    let c = MainUserControl1(DataContext = MainViewModel1())
+    let c = MainUserControl(DataContext = MainViewModel())
 
     do 
         c.ClipToBounds <- true
@@ -23,7 +23,7 @@ type MarginCanvas(getZL) =
             e.Width <- r.Width
             e.SetValue(Canvas.TopProperty, r.Top)
             e.SetValue(Canvas.LeftProperty, r.Left)
-            e.MouseDown.Add(fun _ -> (c.DataContext :?> MainViewModel1).ShowPopup(e.Tag :?> HostIdeActions))
+            e.MouseDown.Add(fun _ -> (c.DataContext :?> MainViewModel).ShowPopup(e.Tag :?> HostIdeActions))
             acc.Add(e) |> ignore
             acc
         c.Canvas.Children.Clear()
